Update to wiki_summary_extract.py
I added comments at the top of the file noting the time and version this module was created.

I also modified the short comment that outlines the purpose of the module.

Lastly, I modified the source code. It now includes a "status" variable that helps the program print a success message at the end of execution.